+++ routingfilter/routing.py
@@ -1,5 +1,6 @@
 import copy
 import json
+import logging
 
 from routingfilter.configfilter import ConfigFilter
 
@@ -8,6 +9,7 @@ class Routing:
 
     def __init__(self):
         self.rules = None
+        self.logger = logging.getLogger(self.__class__.__name__)
 
     def get_rules(self):
         """Return the currently loaded rules. It is mainly used for debugging purposes.
@@ -30,6 +32,7 @@ class Routing:
         :return: A list of dicts containing the matched rules and the outputs in the following format: {"rules": [...], "output": {...}}; an empty list if no rule matched
         """
         if not self.rules:
+            self.logger.error("'rules_list' must be set before evaluating a match!")
             raise ValueError("'rules_list' must be set before evaluating a match!")
         if type_ not in self.rules:
             return []
@@ -73,17 +76,23 @@ class Routing:
                 mr["output"] = mr.pop(type_)
         return matching_rules
 
-    def load_from_dicts(self, rules_list):
+    def load_from_dicts(self, rules_list, validate_rules=True):
         """Load routing configuration from a dictionary. It merges the different rules in list into a single routing rule.
+        It optionally performs some rules validation before accepting them (an exception is raised in case of errors).
 
         :param rules_list: The configuration
         :type rules_list: list[dict]
+        :param validate_rules: Perform rules validation (default=True). It can be disabled to improve performance (unsafe)
+        :type validate_rules: bool
         """
+        self.logger.debug(f"Attempting to load rules_list: {rules_list}")
         rules_list = copy.deepcopy(rules_list)
         if not rules_list:
+            self.logger.warning("An empty rules_list has been passed. Makes sure this is intentional.")
             self.rules = {}
             return self.rules
         if not isinstance(rules_list, list):
+            self.logger.error("'rules_list' must be a list of dicts containing the routing rules!")
             raise ValueError("'rules_list' must be a list of dicts containing the routing rules!")
         merged_rules = rules_list[0]
         for rules in rules_list[1:]:
@@ -96,14 +105,35 @@ class Routing:
                             merged_rules[type_]["rules"][tag] = rules[type_]["rules"][tag]
                 else:
                     merged_rules[type_] = rules[type_]
+        self.logger.debug(f"Merged rules: {merged_rules}")
+        if validate_rules:
+            self._validate_rules(merged_rules)
         self.rules = merged_rules
 
-    def load_from_jsons(self, rules_list):
+    def load_from_jsons(self, rules_list, validate_rules=True):
         """Load routing configuration from JSON data. It merges the different rules in list into a single routing rule.
+        It optionally performs some rules validation before accepting them (an exception is raised in case of errors).
 
         :param rules_list: The json data, which will be parsed into a dict
         :type rules_list: list[str]
+        :param validate_rules: Perform rules validation (default=True). It can be disabled to improve performance (unsafe)
+        :type validate_rules: bool
         """
         if not isinstance(rules_list, str):
+            self.logger.error("'rules_list' must be a list of JSON strings containing the routing rules!")
             raise ValueError("'rules_list' must be a list of JSON strings containing the routing rules!")
-        self.load_from_dicts(json.loads(rules_list))
+        self.load_from_dicts(json.loads(rules_list), validate_rules)
+
+    def _validate_rules(self, rules):
+        """Validate the loaded rules, checking for type mismatch errors.
+
+        :param rules: The merged rules, after loading them from a list of dicts
+        :type rules: dict
+        """
+        self.logger.info("Validating the loaded rules_list")
+        for type_ in rules.keys():
+            for tag_ in rules[type_]["rules"].keys():
+                for filter_output in rules[type_]["rules"][tag_]:
+                    for filter_ in filter_output["filters"]:
+                        config_filter_obj = ConfigFilter(filter_)
+                        config_filter_obj.is_matching({})